// src/app/models/player.model.ts
export class Player {
  first_name?: string;
  last_name?: string;
  position?: string;
  country?: string;
  squad_number?: number;
  league_appearances?: number;
  league_yellow_cards?: number;
  league_red_cards?: number;
  league_mins_played?: number;
  league_shots_on_target?: number;
  league_clean_sheets?: number;
  league_fouls_conceded?: number;
  league_passes?: number;
  league_percentage_of_passes_successful?: number;
  scottish_cup_appearances?: number;
  scottish_cup_yellow_cards?: number;
  scottish_cup_red_cards?: number;
  scottish_cup_mins_played?: number;
  scottish_cup_shots_on_target?: number;
  scottish_cup_clean_sheets?: number;
  scottish_cup_fouls_conceded?: number;
  scottish_cup_passes?: number;
  scottish_cup_percentage_of_passes_successful?: number;
  league_cup_appearances?: number;
  league_cup_yellow_cards?: number;
  league_cup_red_cards?: number;
  league_cup_mins_played?: number;
  league_cup_shots_on_target?: number;
  league_cup_clean_sheets?: number;
  league_cup_fouls_conceded?: number;
  league_cup_passes?: number;
  league_cup_percentage_of_passes_successful?: number;
  europe_appearances?: number;
  europe_yellow_cards?: number;
  europe_red_cards?: number;
  europe_mins_played?: number;
  europe_shots_on_target?: number;
  europe_clean_sheets?: number;
  europe_fouls_conceded?: number;
  europe_passes?: number;
  europe_percentage_of_passes_successful?: number;
  date_of_birth?: Date;
  picture?: string;
  player_id?: number;
}